Refactor contentScripts to use ES6 imports
