+++ manifest.json
@@ -23,18 +23,10 @@
             ],
             "js": [
                 "lib/jquery-2.1.3.min.js",
-                "src/utils/storageManager.js",
-                "src/utils/highlight.js",
-                "src/utils/highlightV3.js",
-                "src/utils/highlightV4.js",
-                "src/utils/highlight_legacy.js",
-                "src/utils/errorManager.js",
-                "src/hoverTools/hoverTools.js",
-                "src/contentScripts/loadHighlights.js",
-                "src/contentScripts/highlighterCursor.js"
+                "contentScript.js"
             ],
             "css": [
-                "src/hoverTools/hoverTools.css"
+                "src/contentScripts/hoverTools/index.css"
             ],
             "all_frames": true
         }
@@ -83,7 +75,7 @@
             "resources": [
                 "images/*.png",
                 "images/*.svg",
-                "src/hoverTools/hoverTools.html"
+                "src/contentScripts/*"
             ],
             "matches": [
                 "<all_urls>"
