+++ package.json
@@ -8,12 +8,14 @@
   "devDependencies": {
     "archiver": "^5.3.0",
     "eslint": "^7.32.0",
+    "eslint-plugin-import": "^2.26.0",
     "readline-sync": "^1.4.10",
     "replace-in-file": "^6.2.0"
   },
   "scripts": {
     "build": "node scripts/build.js",
     "lint": "eslint .",
+    "lint:fix": "eslint --fix .",
     "release": "node scripts/release.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   }
