+++ src/background/actions/getCurrentColor.js
@@ -1,5 +1,6 @@
-import { DEFAULT_COLOR_TITLE } from '../constants.js';
 import getColorOptions from './getColorOptions.js';
+
+import { DEFAULT_COLOR_TITLE } from '../constants.js';
 
 async function getCurrentColor() {
     const { color } = await chrome.storage.sync.get("color");
