+++ src/background/actions/getHighlights.js
@@ -1,7 +1,11 @@
 import { executeInCurrentTab } from '../utils.js';
 
 function getHighlights() {
-    return executeInCurrentTab({ file: 'src/contentScripts/getHighlights.js' });
+    function contentScriptGetHighlights() {
+        return window.getPageHighlights();
+    }
+
+    return executeInCurrentTab({ func: contentScriptGetHighlights });
 }
 
 export default getHighlights;
