+++ src/background/actions/getLostHighlights.js
@@ -3,7 +3,7 @@ import { executeInCurrentTab } from '../utils.js';
 function getLostHighlights() {
     function contentScriptGetLostHighlights() {
         const lostHighlights = [];
-        window.highlighter_lostHighlights.forEach((highlight, index) => lostHighlights.push({ string: highlight?.string, index }));
+        window.getLostHighlights().forEach((highlight, index) => lostHighlights.push({ string: highlight?.string, index }));
         return lostHighlights;
     }
