+++ src/background/actions/index.js
@@ -9,3 +9,4 @@ export { default as changeColor } from './changeColor.js';
 export { default as editColor } from './editColor.js';
 export { default as getCurrentColor } from './getCurrentColor.js';
 export { default as getColorOptions } from './getColorOptions.js';
+export { default as loadPageHighlights } from './loadPageHighlights.js';
