+++ src/background/actions/removeHighlight.js
@@ -5,10 +5,8 @@ function removeHighlight(highlightId) {
     trackEvent('highlight-action', 'remove-highlight');
 
     function contentScriptRemoveHighlight(highlightIndex) {
-        const highlightError = window.highlighter_lostHighlights.get(highlightIndex);
-        clearTimeout(highlightError?.timeout);
-        window.highlighter_lostHighlights.delete(highlightIndex);
-        removeHighlight(highlightIndex, window.location.hostname + window.location.pathname, window.location.pathname);
+        window.removeLostHighlight(highlightIndex);
+        window.removeHighlight(highlightIndex, window.location.hostname + window.location.pathname, window.location.pathname);
     }
 
     executeInCurrentTab({ func: contentScriptRemoveHighlight, args: [highlightId] });
