+++ src/background/actions/showHighlight.js
@@ -5,16 +5,7 @@ function showHighlight(highlightId) {
     trackEvent('highlight-action', 'show-highlight');
 
     function contentScriptShowHighlight(highlightId) { // eslint-disable-line no-shadow
-        const highlightEl = document.querySelector(`[data-highlight-id="${highlightId}"]`);
-        if (highlightEl) {
-            highlightEl.scrollIntoViewIfNeeded(true);
-            const boundingRect = highlightEl.getBoundingClientRect();
-            onHighlightMouseEnterOrClick({
-                'type': 'click',
-                'target': highlightEl,
-                'clientX': boundingRect.left + (boundingRect.width / 2),
-            });
-        }
+        window.showHighlight(highlightId);
     }
 
     executeInCurrentTab({ func: contentScriptShowHighlight, args: [highlightId] });
