+++ src/background/actions/toggleHighlighterCursor.js
@@ -3,7 +3,12 @@ import { executeInCurrentTab } from '../utils.js';
 
 function toggleHighlighterCursor() {
     trackEvent('highlight-action', 'toggle-cursor');
-    executeInCurrentTab({ file: 'src/contentScripts/toggleHighlighterCursor.js' });
+
+    function contentScriptToggleHighlighterCursor() {
+        window.toggleHighlighterCursor();
+    }
+
+    executeInCurrentTab({ func: contentScriptToggleHighlighterCursor });
 }
 
 export default toggleHighlighterCursor;
