+++ src/contentScripts/highlight/constants.js
@@ -1,0 +1,2 @@
+export const HIGHLIGHT_CLASS = 'highlighter--highlighted';
+export const DELETED_CLASS = 'highlighter--deleted';
